chore: move log component to log/component
