+++ frontend/providers/applaunchpad/src/pages/app/detail/logs/components/Header.tsx
